+++ src/accessories/smarthome/smarthome.setup.js
@@ -49,6 +49,7 @@ const Setup = (devices, smarthomeConfig) => {
       if (devices.has(uuid)) {
         logger.warn('Multiple devices are configured with this name. Duplicate devices will be skipped.', device.name);
       } else {
+        logger.debug('New device added!', device.name);
         devices.set(uuid, device);
       }
 
@@ -69,6 +70,7 @@ const Setup = (devices, smarthomeConfig) => {
             tempDevice.name
           );
         } else {
+          logger.debug('New device added!', tempDevice.name);
           devices.set(uuidTemp, tempDevice);
         }
       }
@@ -89,6 +91,7 @@ const Setup = (devices, smarthomeConfig) => {
             windowDevice.name
           );
         } else {
+          logger.debug('New device added!', windowDevice.name);
           devices.set(uuidWindow, windowDevice);
         }
       }
@@ -108,6 +111,7 @@ const Setup = (devices, smarthomeConfig) => {
             openWindowDevice.name
           );
         } else {
+          logger.debug('New device added!', openWindowDevice.name);
           devices.set(uuidWindow, openWindowDevice);
         }
       }
@@ -127,6 +131,7 @@ const Setup = (devices, smarthomeConfig) => {
             blindDevice.name
           );
         } else {
+          logger.debug('New device added!', blindDevice.name);
           devices.set(uuidBlind, blindDevice);
         }
       }